requirements.txt
Flask==3.0.0
openai>=1.30.0
groq>=0.4.0
python-dotenv==1.0.0